Linux port: Fix 64-bit pointer truncation in graphics initialization
Fixed multiple crashes during Instant Action mission launch caused by
32-bit pointer truncation on 64-bit Linux:

- devmgr.cpp: Added Linux workaround in SetRenderTarget() to create
  Direct3D7 interface directly via FF_CreateDirect3D7() instead of
  QueryInterface, which was crashing due to corrupt m_pDD pointer.
  Also skipped AttachDepthBuffer() on Linux since OpenGL manages
  depth buffers automatically.

- context.h/context.cpp: Changed NewImageBuffer() parameter from
  UInt (32-bit unsigned int) to IDirectDrawSurface7* to preserve
  full 64-bit pointer. Also added bounds checking in Stats::Primitive()
  to prevent array underflow when dwType=0.

- render2d.cpp, gmcomposit.cpp: Updated call sites to pass pointer
  directly instead of casting through 32-bit UInt.

- vcock.cpp: Added NULL check for cockpit file to prevent fgets
  crash when file fails to open.

+++ src/graphics/renderer/render2d.cpp
@@ -8,11 +8,11 @@
 #include <cISO646>
 #include <math.h>
 #include "falclib/include/debuggr.h"
-#include "Image.h"
-#include "Device.h"
-#include "Render2D.h"
+#include "image.h"
+#include "device.h"
+#include "render2d.h"
 #include "GraphicsRes.h"
-#include "Tex.h"
+#include "tex.h"
 #include "GraphicsRes.h"
 #include "falclib/include/dispopts.h" //Wombat778 12-12-2003
 
@@ -149,7 +149,7 @@ void Render2D::SetImageBuffer(ImageBuffer *imageBuffer)
 {
     // Remember who our new image buffer is, and tell MPR about the change
     image = imageBuffer;
-    context.NewImageBuffer((DWORD)imageBuffer->targetSurface());
+    context.NewImageBuffer(imageBuffer->targetSurface());
 
     // This shouldn't be required, but _might_ be
     // context.InvalidateState();
